microservices names refactored

+++ search-service/src/main/java/com/stackroute/searchservice/domain/Products.java
@@ -19,7 +19,7 @@ public class Products  {
     private String productFamily;
     private String image;
     private Float price;
-    private int rating;
+    private Float rating;
     private String specifications;
     private String description;
     private Date uploadedOn;
